Updates to graphs & OneRoster handling
Some tidy ups on graph views, have changed relationship graph to bar
chart to make easier to read.

OneRoster merge service uses scan in a loop - scan is a weird function
and never completes in the loop because the db is changing as a result
of what it is doing, so goes recursive.

Have refactored out to a standalone class that can be called via an
/sms endpoint - /sms/merge_ids, which does the work on demand

Call at the end of the seq. shell script and all works fine.